Remove duplicate NATIVE_NULLPTR

+++ app/src/main/kotlin/dev/patrickgold/florisboard/app/ext/ExtensionImportScreen.kt
@@ -48,8 +48,8 @@ import dev.patrickgold.florisboard.app.LocalNavController
 import dev.patrickgold.florisboard.cacheManager
 import dev.patrickgold.florisboard.extensionManager
 import dev.patrickgold.florisboard.ime.keyboard.KeyboardExtension
-import dev.patrickgold.florisboard.ime.nlp.NATIVE_NULLPTR
 import dev.patrickgold.florisboard.ime.theme.ThemeExtension
+import dev.patrickgold.florisboard.lib.NATIVE_NULLPTR
 import dev.patrickgold.florisboard.lib.android.showLongToast
 import dev.patrickgold.florisboard.lib.cache.CacheManager
 import dev.patrickgold.florisboard.lib.compose.FlorisBulletSpacer
@@ -110,14 +110,14 @@ fun ExtensionImportScreen(type: ExtensionImportScreenType, initUuid: String?) = 
                 if (extensionManager.getExtensionById(ext.meta.id)?.sourceRef?.isAssets == true) {
                     R.string.ext__import__file_skip_ext_core
                 } else {
-                    NATIVE_NULLPTR
+                    NATIVE_NULLPTR.toInt()
                 }
             }
             fileInfo.mediaType == FileRegistry.FlexExtension.mediaType -> {
                 R.string.ext__import__file_skip_ext_corrupted
             }
             else -> {
-                NATIVE_NULLPTR
+                NATIVE_NULLPTR.toInt()
             }
         }
     }
@@ -150,7 +150,7 @@ fun ExtensionImportScreen(type: ExtensionImportScreenType, initUuid: String?) = 
             }
             val enabled = remember(importResult) {
                 importResult?.getOrNull()?.takeIf { workspace ->
-                    workspace.inputFileInfos.any { it.skipReason == NATIVE_NULLPTR }
+                    workspace.inputFileInfos.any { it.skipReason == NATIVE_NULLPTR.toInt() }
                 } != null
             }
             ButtonBarButton(
@@ -160,7 +160,7 @@ fun ExtensionImportScreen(type: ExtensionImportScreenType, initUuid: String?) = 
                 val workspace = importResult!!.getOrThrow()
                 runCatching {
                     for (fileInfo in workspace.inputFileInfos) {
-                        if (fileInfo.skipReason != NATIVE_NULLPTR) {
+                        if (fileInfo.skipReason != NATIVE_NULLPTR.toInt()) {
                             continue
                         }
                         val ext = fileInfo.ext
@@ -304,7 +304,7 @@ private fun FileInfoView(
                     )
                 }
             }
-            if (fileInfo.skipReason != NATIVE_NULLPTR) {
+            if (fileInfo.skipReason != NATIVE_NULLPTR.toInt()) {
                 Box(modifier = Modifier
                     .fillMaxWidth()
                     .height(19.dp)
