+++ app/src/main/kotlin/dev/patrickgold/florisboard/lib/cache/CacheManager.kt
@@ -26,8 +26,8 @@ import androidx.compose.runtime.setValue
 import dev.patrickgold.florisboard.app.ext.EditorAction
 import dev.patrickgold.florisboard.app.settings.advanced.Backup
 import dev.patrickgold.florisboard.appContext
-import dev.patrickgold.florisboard.ime.nlp.NATIVE_NULLPTR
 import dev.patrickgold.florisboard.ime.theme.ThemeExtensionEditor
+import dev.patrickgold.florisboard.lib.NATIVE_NULLPTR
 import dev.patrickgold.florisboard.lib.android.query
 import dev.patrickgold.florisboard.lib.android.readToFile
 import dev.patrickgold.florisboard.lib.ext.Extension
@@ -235,6 +235,6 @@ class CacheManager(context: Context) {
         val mediaType: String?,
         val size: Long,
         val ext: Extension?,
-        var skipReason: Int = NATIVE_NULLPTR,
+        var skipReason: Int = NATIVE_NULLPTR.toInt(),
     )
 }
